reorganized structure of the supercharger.optimize subpackage.
