+++ python/src/supercharger/optimize/__init__.py
@@ -1,0 +1,15 @@
+"""
+The public interface of the supercharger 'optimize' package.
+"""
+from ._constraints import (
+    ConstraintData,
+)
+
+from ._optimizer import (
+    NonlinearOptimizer
+)
+
+__all__ = [
+    'ConstraintData',
+    'NonlinearOptimizer'
+]
